Implemented Reusability Concepts on Homepage and subsections of the Homepage.
